fixed line endings & naming of files
